pridavane postupne zakladne principy pre funkcnosti simulacie

+++ src/Simulation/Events/CustomerArrived.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class CustomerArrived extends Event{
-    public CustomerArrived(int time, Simulator simulationCore) {
+    public CustomerArrived(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
