+++ src/Simulation/Events/HairstyleBeginning.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class HairstyleBeginning extends Event{
-    public HairstyleBeginning(int time, Simulator simulationCore) {
+    public HairstyleBeginning(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
