+++ src/Simulation/Events/HairstyleEnd.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class HairstyleEnd extends Event{
-    public HairstyleEnd(int time, Simulator simulationCore) {
+    public HairstyleEnd(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
