+++ src/Simulation/Events/MakeupBeginning.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class MakeupBeginning extends Event{
-    public MakeupBeginning(int time, Simulator simulationCore) {
+    public MakeupBeginning(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
