+++ src/Simulation/Events/MakeupEnd.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class MakeupEnd extends Event{
-    public MakeupEnd(int time, Simulator simulationCore) {
+    public MakeupEnd(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
