+++ src/Simulation/Events/PaymentBeginning.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class PaymentBeginning extends Event{
-    public PaymentBeginning(int time, Simulator simulationCore) {
+    public PaymentBeginning(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
