+++ src/Simulation/Events/PaymentEnd.java
@@ -3,7 +3,7 @@ package Simulation.Events;
 import Simulation.Simulator;
 
 public class PaymentEnd extends Event{
-    public PaymentEnd(int time, Simulator simulationCore) {
+    public PaymentEnd(double time, Simulator simulationCore) {
         super(time, simulationCore);
     }
